instrinsics and sorted videofiles READ DETAILS
calibration turned out to be a slow process as such i've only been able
to get the intrinsic camera data, it can be found in the assignment 1
folder.

Due to reasons i cannot upload video files to git as such i've uploaded
the sorted files to a cloud service, following is the link

http://sdrv.ms/1aHkOOA

just download and unzip. inform me if there is something wrong with the
link