Added Clustering class that does initial clustering.
The first frame is now #195, in which the 4-person video is kind of well
separated. The voxels active in this frame are clustered in (x,y)
position, this is shown visually with a top-down drawing of the voxels
with colors based on the cluster. <This version does not work perfectly
yet, but at least it shows something!>