Online DONE, though not entirely perfect it's pretty good (SEE DESCRIPTION)
A couple of notes
1: this thing takes a bit to do it's output. it's does not live show the
results in realtime. I've added a progress indicator in the console
output.

2: just consider the function runtracking() to be a blackbox it should
work with any amount of clusters and any video input. Note i did not
have a chance to write a lot of comments so i'd have to urge you to not
change things here unless you have to.

note that the results are really good with the 2 persons video
i did not test the 4p

do feel free to contact me on skype.